+++ packages/renative-template-hello-world/src/app.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { createApp } from 'renative';
+import { createApp, registerFocusManger, registerServiceWorker } from 'renative';
 import { navStructure } from './nav';
 import ScreenHome from './screenHome';
 import ScreenMyPage from './screenMyPage';
@@ -7,6 +7,9 @@ import ScreenModal from './screenModal';
 import Menu from './menu';
 
 import '../platformAssets/runtime/fontManager';
+
+registerFocusManger({ focused: 'border: 5px solid #62DBFB; border-radius:5px;' });
+registerServiceWorker();
 
 let AppContainer;