+++ packages/rnv/platformTemplates/tizen/_privateConfig/webpack.config.dev.js
@@ -5,7 +5,6 @@ const config = {
     currentDir: __dirname,
     metaTags: { viewport: 'width=device-width, initial-scale=1, shrink-to-fit=no' },
     environment: 'development',
-    extensions: ['tizen', 'smarttv', 'web'],
     customScripts: [],
     devServerHost: '0.0.0.0',
     baseUrl: '',
@@ -17,7 +16,7 @@ const plugins = [C.Plugins.webpack, C.Plugins.html, C.Plugins.harddisk];
 if (config.analyzer) plugins.push(C.Plugins.analyzer);
 
 module.exports = {
-    entry: C.entry,
+    entry: { raf: 'raf/polyfill', ...C.entry },
     devServer: C.devServer,
     output: C.output,
     optimization: {