+++ packages/rnv/platformTemplates/tizenmobile/_privateConfig/webpack.config.dev.js
@@ -5,7 +5,6 @@ const config = {
     currentDir: __dirname,
     metaTags: { viewport: 'content="width=device-width, initial-scale=1.0, maximum-scale=1.0"' },
     environment: 'development',
-    extensions: ['tizenmobile', 'web'],
     customScripts: [],
     devServerHost: '0.0.0.0',
     baseUrl: '',