+++ packages/rnv/src/cli/index.js
@@ -60,7 +60,8 @@ ${chalk.bold.white('OPTIONS:')}
 
 const COMMANDS = {
     start: {
-        fn: rnvStart
+        fn: rnvStart,
+        platforms: SUPPORTED_PLATFORMS
     },
     run: {
         desc: 'Run your app on target device or emulator',
@@ -296,12 +297,12 @@ const _execute = async (c, cmdFn, cmd, command, subCommand) => {
     let subCmd = '';
     if (subCommand) {
         subCmd = `:${c.subCommand}`;
-        const requiredPlatforms = cmd.subCommands[c.subCommand]?.platforms;
+        const requiredPlatforms = cmd.subCommands?.[c.subCommand]?.platforms;
         if (requiredPlatforms && !requiredPlatforms.includes(c.platform)) {
             await _handleUnknownPlatform(c, requiredPlatforms);
             return;
         }
-        const requiredParams = cmd.subCommands[c.subCommand]?.requiredParams;
+        const requiredParams = cmd.subCommands?.[c.subCommand]?.requiredParams;
         if (requiredParams) {
             for (let i = 0; i < requiredParams.length; i++) {
                 const requiredParam = requiredParams[i];