+++ packages/rnv/src/platformTools/apple/index.js
@@ -49,7 +49,9 @@ const runPod = (c, command, cwd, rejectOnFail = false) => new Promise((resolve, 
             cwd,
             evn: process.env,
         })
-            .then(() => resolve())
+            .then(() => {
+                resolve();
+            })
             .catch((e) => {
                 if (rejectOnFail) {
                     logWarning(e);
@@ -186,7 +188,7 @@ const _runXcodeProject = async (c, platform, target) => {
 
     if (p) {
         const allowProvisioningUpdates = getConfigProp(c, platform, 'allowProvisioningUpdates', true);
-        if (allowProvisioningUpdates) p.push('--allowProvisioningUpdates');
+        // if (allowProvisioningUpdates) p.push('--allowProvisioningUpdates');
 
         if (bundleAssets) {
             return packageBundleForXcode(c, platform, bundleIsDev).then(() => executeAsync(c, `react-native ${p.join(' ')}`));
@@ -205,6 +207,14 @@ const _checkLockAndExec = (c, p) => executeAsync(c, `react-native ${p}`)
         }
         return Promise.reject(e);
     });
+
+const composeXcodeArgsFromCLI = (string) => {
+    const spacesReplaced = string.replace(/\s(?=(?:[^'"`]*(['"`])[^'"`]*\1)*[^'"`]*$)/g, '&&&'); // replaces spaces outside quotes with &&& for easy split
+    const keysAndValues = spacesReplaced.split('&&&');
+    const unescapedValues = keysAndValues.map(s => s.replace(/\'/g, '').replace(/"/g, '').replace(/\\/g, '')); // removes all quotes or backslashes
+
+    return unescapedValues;
+};
 
 const archiveXcodeProject = (c, platform) => {
     logTask(`archiveXcodeProject:${platform}`);
@@ -235,21 +245,25 @@ const archiveXcodeProject = (c, platform) => {
     const p = [];
 
     if (!ps.includes('-workspace')) {
-        p.push(`-workspace ${appPath}/${appFolderName}.xcworkspace`);
+        p.push('-workspace');
+        p.push(`${appPath}/${appFolderName}.xcworkspace`);
     }
     if (!ps.includes('-scheme')) {
-        p.push(`-scheme ${scheme}`);
+        p.push('-scheme');
+        p.push(scheme);
     }
     if (!ps.includes('-sdk')) {
         p.push('-sdk');
         p.push(...sdkArr);
     }
     if (!ps.includes('-configuration')) {
-        p.push(`-configuration ${runScheme}`);
+        p.push('-configuration');
+        p.push(runScheme);
     }
     p.push('archive');
     if (!ps.includes('-archivePath')) {
-        p.push(`-archivePath ${exportPathArchive}`);
+        p.push('-archivePath');
+        p.push(exportPathArchive);
     }
 
     if (allowProvisioningUpdates && !ps.includes('-allowProvisioningUpdates')) p.push('-allowProvisioningUpdates');
@@ -266,7 +280,12 @@ const archiveXcodeProject = (c, platform) => {
                 logSuccess(`Your Archive is located in ${chalk.white(exportPath)} .`);
             });
     }
-    return executeAsync(c, `xcodebuild ${ps} ${p.join(' ')}`)
+
+    const args = ps !== '' ? [...composeXcodeArgsFromCLI(ps), ...p] : p;
+
+    logDebug('xcodebuild args', args);
+
+    return executeAsync('xcodebuild', { rawCommand: { args } })
         .then(() => {
             logSuccess(`Your Archive is located in ${chalk.white(exportPath)} .`);
         });
@@ -436,17 +455,19 @@ const configureXcodeProject = (c, platform, ip, port) => new Promise((resolve, r
     const forceUpdate = !fs.existsSync(path.join(appFolder, 'Podfile.lock')) || c.program.update;
     copyAssetsFolder(c, platform)
         .then(() => copyAppleAssets(c, platform, appFolderName))
-        .then(() => copyBuildsFolder(c, platform))
         .then(() => parseAppDelegate(c, platform, appFolder, appFolderName, bundleAssets, bundlerIp, port))
         .then(() => parseExportOptionsPlist(c, platform))
         .then(() => parseXcscheme(c, platform))
         .then(() => parsePodFile(c, platform))
         .then(() => parseEntitlementsPlist(c, platform))
         .then(() => parseInfoPlist(c, platform))
+        .then(() => copyBuildsFolder(c, platform))
         .then(() => {
             runPod(c, forceUpdate ? 'update' : 'install', getAppFolder(c, platform), true)
                 .then(() => parseXcodeProject(c, platform))
-                .then(() => resolve())
+                .then(() => {
+                    resolve();
+                })
                 .catch((e) => {
                     if (!c.program.update) {
                         if (e && e.toString) {
@@ -455,6 +476,7 @@ const configureXcodeProject = (c, platform, ip, port) => new Promise((resolve, r
                                 s.includes('No provisionProfileSpecifier configured')
                               || s.includes('TypeError:')
                               || s.includes('ReferenceError:')
+                              || s.includes('find gem cocoapods')
                             ) {
                                 reject(e);
                             }