+++ packages/rnv/src/platformTools/apple/xcodeParser.js
@@ -1,6 +1,5 @@
 import path from 'path';
 import fs from 'fs';
-import chalk from 'chalk';
 import {
     logTask,
     logError,
@@ -20,18 +19,15 @@ import {
     logSuccess,
     getBuildsFolder,
 } from '../../common';
-import { copyBuildsFolder } from '../../projectTools/projectParser';
 import { inquirerPrompt } from '../../systemTools/prompt';
 import { IOS, TVOS } from '../../constants';
-import { getMergedPlugin, parsePlugins } from '../../pluginTools';
+import { parsePlugins } from '../../pluginTools';
 import { getAppFolderName } from './index';
 import { parseProvisioningProfiles } from './provisionParser';
-import { copyFolderContentsRecursiveSync, copyFileSync, mkdirSync, readObjectSync, mergeObjects, writeObjectSync } from '../../systemTools/fileutils';
-
-const xcode = require('xcode');
-
+import { writeObjectSync } from '../../systemTools/fileutils';
 
 export const parseXcodeProject = async (c, platform) => {
+    logTask('parseXcodeProject');
     // PROJECT
     c.runtime.xcodeProj = {};
     c.runtime.xcodeProj.provisioningStyle = getConfigProp(c, platform, 'provisioningStyle', 'Automatic');
@@ -77,7 +73,10 @@ export const parseXcodeProject = async (c, platform) => {
     await _parseXcodeProject(c, platform);
 };
 
-const _parseXcodeProject = (c, platform, config) => new Promise((resolve, reject) => {
+const _parseXcodeProject = (c, platform) => new Promise((resolve, reject) => {
+    logTask('_parseXcodeProject');
+    // eslint-disable-next-line global-require, import/no-dynamic-require
+    const xcode = require(`${c.paths.project.nodeModulesDir}/xcode`);
     const appFolder = getAppFolder(c, platform);
     const appFolderName = getAppFolderName(c, platform);
     const projectPath = path.join(appFolder, `${appFolderName}.xcodeproj/project.pbxproj`);
@@ -124,7 +123,6 @@ const _parseXcodeProject = (c, platform, config) => new Promise((resolve, reject
             }
         }
 
-
         if (systemCapabilities) {
             const sysCapObj = {};
             for (const sk in systemCapabilities) {
@@ -134,12 +132,10 @@ const _parseXcodeProject = (c, platform, config) => new Promise((resolve, reject
             // const var1 = xcodeProj.getFirstProject().firstProject.attributes.TargetAttributes['200132EF1F6BF9CF00450340'];
             xcodeProj.addTargetAttribute('SystemCapabilities', sysCapObj);
         }
-
         // FONTS
         c.pluginConfigiOS.embeddedFontSources.forEach((v) => {
             xcodeProj.addResourceFile(v);
         });
-
 
         // PLUGINS
         parsePlugins(c, platform, (plugin, pluginPlat, key) => {
@@ -196,28 +192,7 @@ const _parseXcodeProject = (c, platform, config) => new Promise((resolve, reject
                 }
             }
         });
-
         fs.writeFileSync(projectPath, xcodeProj.writeSync());
         resolve();
     });
 });
-
-
-// export const parseXcodeProjec2() => new Promise((resolve, reject) => {
-// const projectPath = path.join(appFolder, `${appFolderName}.xcodeproj/project.pbxproj`);
-// const xcodeProj = xcode.project(projectPath);
-// xcodeProj.parse(() => {
-//     const appId = getAppId(c, platform);
-//     if (teamID) {
-//         xcodeProj.updateBuildProperty('DEVELOPMENT_TEAM', teamID);
-//     } else {
-//         xcodeProj.updateBuildProperty('DEVELOPMENT_TEAM', '""');
-//     }
-//
-//     xcodeProj.addTargetAttribute('ProvisioningStyle', provisioningStyle);
-//     xcodeProj.addBuildProperty('CODE_SIGN_STYLE', provisioningStyle);
-//     xcodeProj.updateBuildProperty('PRODUCT_BUNDLE_IDENTIFIER', appId);
-//
-//     resolve();
-// });
-// })