+++ packages/rnv/src/platformTools/electron/index.js
@@ -33,24 +33,23 @@ import {
     waitForWebpack
 } from '../../common';
 import { copyBuildsFolder, copyAssetsFolder } from '../../projectTools/projectParser';
-import { MACOS } from '../../constants';
-import { buildWeb, runWeb, runWebDevServer } from '../web';
+import { MACOS, WINDOWS } from '../../constants';
+import { buildWeb, runWeb } from '../web';
 import {
     cleanFolder, copyFolderContentsRecursiveSync, copyFolderRecursiveSync,
-    copyFileSync, mkdirSync, writeObjectSync, readObjectSync
+    copyFileSync, mkdirSync, writeObjectSync, readObjectSync, removeDirs, removeDirsSync
 } from '../../systemTools/fileutils';
 
 const isRunningOnWindows = process.platform === 'win32';
 
-const configureElectronProject = (c, platform) => new Promise((resolve, reject) => {
+const configureElectronProject = async (c, platform) => {
     logTask(`configureElectronProject:${platform}`);
 
-    copyAssetsFolder(c, platform, platform === MACOS ? _generateICNS : null)
-        .then(() => copyBuildsFolder(c, platform))
-        .then(() => configureProject(c, platform))
-        .then(() => resolve())
-        .catch(e => reject(e));
-});
+    await copyAssetsFolder(c, platform, platform === MACOS ? _generateICNS : null);
+    await configureProject(c, platform);
+    return copyBuildsFolder(c, platform);
+};
+const merge = require('deepmerge');
 
 const configureProject = (c, platform) => new Promise((resolve, reject) => {
     logTask(`configureProject:${platform}`);
@@ -78,7 +77,7 @@ const configureProject = (c, platform) => new Promise((resolve, reject) => {
     const packageJson = readObjectSync(pkgJson);
 
     packageJson.name = `${getAppConfigId(c, platform)}-${platform}`;
-    packageJson.productName = `${getAppTitle(c, platform)} - ${platform}`;
+    packageJson.productName = `${getAppTitle(c, platform)}`;
     packageJson.version = `${getAppVersion(c, platform)}`;
     packageJson.description = `${getAppDescription(c, platform)}`;
     packageJson.author = getAppAuthor(c, platform);
@@ -104,14 +103,28 @@ const configureProject = (c, platform) => new Promise((resolve, reject) => {
         );
     }
 
-    const electronConfig = {
+    const macConfig = {};
+    if (platform === MACOS) {
+        macConfig.mac = {
+            entitlements: path.join(appFolder, 'entitlements.mac.plist'),
+            entitlementsInherit: path.join(appFolder, 'entitlements.mac.plist')
+        };
+    }
+
+    let electronConfig = merge({
         appId,
         directories: {
             app: appFolder,
             buildResources: path.join(appFolder, 'resources'),
             output: path.join(appFolder, 'build/release')
-        }
-    };
+        },
+    }, macConfig);
+
+    const electronConfigExt = getConfigProp(c, platform, 'electronConfig');
+
+    if (electronConfigExt) {
+        electronConfig = merge(electronConfig, electronConfigExt);
+    }
     writeObjectSync(electronConfigPath, electronConfig);
 
 
@@ -124,18 +137,21 @@ const buildElectron = (c, platform) => {
     return buildWeb(c, platform);
 };
 
-const exportElectron = (c, platform) => new Promise((resolve, reject) => {
+const exportElectron = async (c, platform) => {
     logTask(`exportElectron:${platform}`);
     const { maxErrorLength } = c.program;
-
     const appFolder = getAppFolder(c, platform);
-    executeAsync(c, `npx electron-builder --config ${path.join(appFolder, 'electronConfig.json')}`)
-        .then(() => {
-            logSuccess(`Your Exported App is located in ${chalk.white(path.join(appFolder, 'build/release'))} .`);
-            resolve();
-        })
-        .catch(e => reject(e));
-});
+    const buildPath = path.join(appFolder, 'build');
+
+    if (fs.existsSync(buildPath)) {
+        console.log(`removing old build ${buildPath}`);
+        await removeDirs([buildPath]);
+    }
+
+    await executeAsync(c, `npx electron-builder --config ${path.join(appFolder, 'electronConfig.json')}`);
+
+    logSuccess(`Your Exported App is located in ${chalk.white(path.join(appFolder, 'build/release'))} .`);
+};
 
 const runElectron = async (c, platform, port) => {
     logTask(`runElectron:${platform}`);
@@ -175,7 +191,7 @@ const _runElectronSimulator = (c, platform) => new Promise((resolve, reject) => 
     const appFolder = getAppFolder(c, platform);
     const elc = resolveNodeModulePath(c, 'electron/cli.js');
 
-    const child = spawn('node', [elc, appFolder], {
+    const child = spawn('node', [elc, path.join(appFolder, '/main.js')], {
         detached: true,
         env: process.env,
         stdio: 'inherit',
@@ -197,7 +213,7 @@ const runElectronDevServer = (c, platform, port) => new Promise((resolve, reject
         path.join(appFolder, 'webpack.config.js')
     );
 
-    runWebDevServer(c, platform, port)
+    runWeb(c, platform, port)
         .then(() => resolve())
         .catch(e => reject(e));
 });