+++ packages/rnv/src/platformTools/firefox/index.js
@@ -69,17 +69,15 @@ const launchKaiOSSimulator = (c, name) => new Promise((resolve, reject) => {
     });
 });
 
-const configureKaiOSProject = (c, platform) => new Promise((resolve, reject) => {
+const configureKaiOSProject = async (c, platform) => {
     logTask('configureKaiOSProject');
 
-    if (!isPlatformActive(c, platform, resolve)) return;
+    if (!isPlatformActive(c, platform)) return;
 
-    copyAssetsFolder(c, platform)
-        .then(() => copyBuildsFolder(c, platform))
-        .then(() => configureProject(c, platform))
-        .then(() => resolve())
-        .catch(e => reject(e));
-});
+    await copyAssetsFolder(c, platform);
+    await configureProject(c, platform);
+    return copyBuildsFolder(c, platform);
+};
 
 const configureProject = (c, platform) => new Promise((resolve, reject) => {
     logTask(`configureProject:${platform}`);